Replace None with '' for added/deleted files
Let mylist = ['b', 'a', 'c', None].
In Python2 `sorted(mylist)` returns `[None, 'a', 'b', 'c'].
In Python3 `sorted(mylist)` errors out because you can't compare a
`None` to strings.
Because of this difference I chose to replace `None` with `''` for files
which have been added or deleted. It still makes sense logically because
the path to the file is non-existent so `''` and the name of the file is
`''`. I'm passing the tests, but the browser page shows no diff. I have
no experience using flask or any other web frameworks, so I don't know
how to find what's causing the blank web page.